% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-filter.R
\name{dust_filter_copy}
\alias{dust_filter_copy}
\title{Create copy of filter}
\usage{
dust_filter_copy(filter, seed = NULL)
}
\arguments{
\item{filter}{A \code{dust_filter} object, created by
\link{dust_filter_create}}

\item{seed}{The seed for the filter (see \link{dust_filter_create})}
}
\value{
A new \code{dust_filter} object
}
\description{
Create an independent copy of a filter.  The new filter is
decoupled from the random number streams of the parent filter.  It
is also decoupled from the \emph{state size} of the parent filter, so
you can use this to create a new filter where the system is
fundamentally different but everything else is the same.
}
