% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dust_filter_data}
\alias{dust_filter_data}
\title{Prepare data}
\usage{
dust_filter_data(data, time = NULL, group = NULL, n_groups = NULL)
}
\arguments{
\item{data}{A data.frame containing time and data to fit.  By
default we expect a column \code{time} (or one with the name given as
the argument \code{time}) and one or more columns of data to fit to.}

\item{time}{Optional name of a column within \code{data} to use for
time.}

\item{group}{Optional name of a column within \code{data} to use for
groups}

\item{n_groups}{Optional number of parameter groups that this
dataset applies to.}
}
\value{
A data.frame, with the addition of the class attribute
\code{dust_filter_data}; once created you should not modify this
object.
}
\description{
Prepare data for use with \link{dust_unfilter_create} or
\link{dust_filter_create}.  You do not have to use this function if you
name your data.frame with our standard column names (i.e., \code{time}
column containing the time) as it will be called within the filter
functions directly.  However, you can use this to validate your
data separately or to use different columns than the defaults.
}
