% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-filter.R
\name{dust_filter_last_history}
\alias{dust_filter_last_history}
\title{Fetch last filter history}
\usage{
dust_filter_last_history(filter, index_group = NULL)
}
\arguments{
\item{filter}{A \code{dust_filter} object, created by
\link{dust_filter_create}}

\item{index_group}{An optional vector of group indices to run the
filter for.  You can use this to run a subset of possible
groups, though at the moment this will likely crash.}
}
\value{
An array
}
\description{
Fetch the last history created by running a filter.  This
errors if the last call to \link{dust_filter_run} did not use
\code{save_history = TRUE}.
}
