% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-filter.R
\name{dust_filter_last_state}
\alias{dust_filter_last_state}
\title{Get filter state}
\usage{
dust_filter_last_state(
  filter,
  index_group = NULL,
  select_random_particle = FALSE
)
}
\arguments{
\item{filter}{A \code{dust_filter} object, created by
\link{dust_filter_create}}

\item{index_group}{An optional vector of group indices to run the
filter for.  You can use this to run a subset of possible
groups, once the filter is initialised (this argument must be
\code{NULL} on the \strong{first} call).}

\item{select_random_particle}{Logical, indicating if we should
return a history for one randomly selected particle (rather than
the entire history).  If this is \code{TRUE}, the particle will be
selected independently for each group, if the filter is grouped.
This option is intended to help select a representative
trajectory during an MCMC.  When \code{TRUE}, we drop the \code{particle}
dimension of the return value.}
}
\value{
An array.  If ungrouped this will have dimensions \code{state}
x \code{particle}, and if grouped then \code{state} x \code{particle} x
\code{group}.  If \code{select_random_particle = TRUE}, the second
(particle) dimension will be dropped.  This is the same as the
state returned by \link{dust_filter_last_history} without the time
dimension but also without any state index applied (i.e., we
always return all state).
}
\description{
Get the last state from a filter.
}
