% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_filter_run}
\alias{dust_filter_run}
\title{Run particle filter}
\usage{
dust_filter_run(filter, pars = NULL, initial = NULL)
}
\arguments{
\item{filter}{A \code{dust_filter} object, created by
\link{dust_filter_create}}

\item{pars}{Optional parameters to run the filter with.  If not
provided, parameters are not updated}

\item{initial}{Optional initial conditions, as a matrix (state x
particle) or 3d array (state x particle x group).  If not
provided, the model initial conditions are used.}
}
\value{
A vector of likelihood values, with as many elements as
there groups.
}
\description{
Run particle filter
}
