% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-likelihood.R
\name{dust_likelihood_copy}
\alias{dust_likelihood_copy}
\title{Create copy of a dust likelihood object}
\usage{
dust_likelihood_copy(obj, seed = NULL)
}
\arguments{
\item{obj}{A \code{dust_filter} object, created by
\link{dust_filter_create} or a \code{dust_unfilter} object created by
\link{dust_unfilter_create}}

\item{seed}{The seed for the particle filter (see
\link{dust_filter_create})}
}
\value{
A new \code{dust_likelihood} object
}
\description{
Create an independent copy of a likelihood object.  The new object
is decoupled from the random number streams of the parent object.
It is also decoupled from the \emph{state size} of the parent object,
so you can use this to create a new object where the system is
fundamentally different but everything else is the same.
}
