% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-likelihood.R
\name{dust_likelihood_last_gradient}
\alias{dust_likelihood_last_gradient}
\title{Fetch last likelihood gradient}
\usage{
dust_likelihood_last_gradient(obj)
}
\arguments{
\item{obj}{A \code{dust_filter} object, created by
\link{dust_filter_create} or a \code{dust_unfilter} object created by
\link{dust_unfilter_create}}
}
\value{
A vector (if ungrouped) or a matrix (if grouped).
}
\description{
Fetch the last gradient created by running an likelihood.  This
errors if the last call to \link{dust_likelihood_run} did not use
\code{adjoint = TRUE}.  The first time you call this (after a
particular set of parameters) it will trigger running the reverse
model.
}
