% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-likelihood.R
\name{dust_likelihood_last_history}
\alias{dust_likelihood_last_history}
\title{Fetch last likelihood history}
\usage{
dust_likelihood_last_history(
  obj,
  index_state = NULL,
  index_group = NULL,
  select_random_particle = FALSE
)
}
\arguments{
\item{obj}{A \code{dust_filter} object, created by
\link{dust_filter_create} or a \code{dust_unfilter} object created by
\link{dust_unfilter_create}}

\item{index_state}{An optional vector of state indices to save
where \code{save_history} is \code{TRUE}.  If omitted all state is saved
and if \code{save_history = FALSE} this has no effect.}

\item{index_group}{An optional vector of group indices to run the
calculation for.  You can use this to run a subset of possible
groups, once \code{obj} is initialised (this argument must be \code{NULL}
on the \strong{first} call).}

\item{select_random_particle}{Logical, indicating if we should
return a history for one randomly selected particle (rather than
the entire history).  If this is \code{TRUE}, the particle will be
selected independently for each group, if the object is grouped.
This option is intended to help select a representative
trajectory during an MCMC.  When \code{TRUE}, we drop the \code{particle}
dimension of the return value.}
}
\value{
An array.  If ungrouped this will have dimensions \code{state}
x \code{particle} x \code{time}, and if grouped then \code{state} x \code{particle}
x \code{group} x \code{time}.  If \code{select_random_particle = TRUE}, the
second (particle) dimension will be dropped.
}
\description{
Fetch the last history created by running a likelihood.  This
errors if the last call to \link{dust_likelihood_run} did not use
\code{save_history = TRUE}.
}
