% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-likelihood.R
\name{dust_likelihood_last_trajectories}
\alias{dust_likelihood_last_trajectories}
\title{Fetch last likelihood trajectories}
\usage{
dust_likelihood_last_trajectories(obj, select_random_particle = FALSE)
}
\arguments{
\item{obj}{A \code{dust_filter} object, created by
\link{dust_filter_create} or a \code{dust_unfilter} object created by
\link{dust_unfilter_create}}

\item{select_random_particle}{Logical, indicating if we should
return a history for one randomly selected particle (rather than
the entire set of trajectories over all particles).  If this is
\code{TRUE}, the particle will be selected independently for each
group, if the object is grouped.  This option is intended to
help select a representative trajectory during an MCMC.  When
\code{TRUE}, we drop the \code{particle} dimension of the return value.}
}
\value{
An array.  If ungrouped this will have dimensions \code{state}
x \code{particle} x \code{time}, and if grouped then \code{state} x \code{particle}
x \code{group} x \code{time}.  If \code{select_random_particle = TRUE}, the
second (particle) dimension will be dropped.
}
\description{
Fetch the last trajectories created by running a likelihood.  This
errors if the last call to \link{dust_likelihood_run} did not use
\code{save_trajectories = TRUE}.  We return the states and groups that
were run via the \code{index_state} and \code{index_group} arguments to
\link{dust_likelihood_run}.
}
