% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_compare_data}
\alias{dust_model_compare_data}
\title{Compare model state against data}
\usage{
dust_model_compare_data(model, data)
}
\arguments{
\item{model}{A \code{dust_model} object}

\item{data}{The data to compare against. If the model is ungrouped
then \code{data} is a list with elements corresponding to whatever
your model requires.  If your model is grouped, this should be a
list with as many elements as your model has groups, with each
element corresponding to the data your model requires.}
}
\value{
A numeric vector with as many elements as your model has
groups, corresponding to the log-likelihood of the data for each
group.
}
\description{
Compare current model state against data.  This is only supported
for models that have 'compare_data' support (i.e., the model
definition includes a \code{compare_data} method).  The current state
in the model (\link{dust_model_state}) is compared against the data
provided as \code{data}.
}
