% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_rng_state}
\alias{dust_model_rng_state}
\title{Fetch rng state}
\usage{
dust_model_rng_state(model)
}
\arguments{
\item{model}{A \code{dust_model} object}
}
\value{
A raw vector, this could be quite long.
}
\description{
Fetch the random number generator (RNG) state from the model.
}
\seealso{
You can pass the state you get back from this function as
the seed object to \code{dust_model_create}.  In future,
\code{dust_model_set_rng_state} will also let you set state into an
existing model.
}
