% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_run_steps}
\alias{dust_model_run_steps}
\alias{dust_model_run_to_time}
\title{Run model}
\usage{
dust_model_run_steps(model, steps)

dust_model_run_to_time(model, time)
}
\arguments{
\item{model}{A \code{dust_model} object}

\item{steps}{The number of steps to run forward}

\item{time}{Time to run to}
}
\value{
Nothing, called for side effects only
}
\description{
Run a model, advancing time and the state by repeatedly running
its \code{update} method.  You can advance a model either a fixed
(positive) number of steps, or up to a time (which must be in the
future).
}
