% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_set_state}
\alias{dust_model_set_state}
\title{Set model state}
\usage{
dust_model_set_state(model, state)
}
\arguments{
\item{model}{A \code{dust_model} object}

\item{state}{A matrix or array of state.  If ungrouped, the
dimension order expected is state x particle.  If grouped the
order is state x particle x group.}
}
\value{
Nothing, called for side effects only
}
\description{
Set model state.  Takes a multidimensional array (2- or 3d
depending on if the model is grouped or not).  Dimensions of
length 1 will be recycled as appropriate.
}
