% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_set_state_initial}
\alias{dust_model_set_state_initial}
\title{Set model state to initial conditions}
\usage{
dust_model_set_state_initial(model)
}
\arguments{
\item{model}{A \code{dust_model} object}
}
\value{
Nothing, called for side effects only
}
\description{
Set model state from a model's initial conditions.  This may depend
on the current time.
}
