% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_simulate}
\alias{dust_model_simulate}
\title{Simulate model}
\usage{
dust_model_simulate(model, times, index = NULL)
}
\arguments{
\item{model}{A \code{dust_model} object}

\item{times}{A vector of times.  They must be increasing, and the
first time must be no less than the current model time
(as reported by \link{dust_model_time})}

\item{index}{An optional index of states to extract.  If given,
then we subset the model state on return.  You can use this to
return fewer model states than the model ran with, to reorder
states, or to name them on exit (names present on the index will
be copied into the rownames of the returned array).}
}
\value{
An array with 3 dimensions (state x particle x time) or 4
dimensions (state x particle x group x time) for a grouped
model.
}
\description{
Simulate a model over a series of times, returning an array of
output.  This output can be quite large, so you may filter states
according to some index.
}
