% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_state}
\alias{dust_model_state}
\title{Extract model state}
\usage{
dust_model_state(model)
}
\arguments{
\item{model}{A \code{dust_model} object}
}
\value{
An array of model state.  If your model is ungrouped, then
this has two dimensions (state, particle).  If grouped, this has
three dimensions (state, particle, group)
}
\description{
Extract model state
}
\seealso{
\code{\link[=dust_model_set_state]{dust_model_set_state()}} for setting state and
\code{\link[=dust_model_set_state_initial]{dust_model_set_state_initial()}} for setting state to the
model-specific initial conditions.
}
