% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_time}
\alias{dust_model_time}
\title{Fetch model time}
\usage{
dust_model_time(model)
}
\arguments{
\item{model}{A \code{dust_model} object}
}
\value{
A single numeric value
}
\description{
Fetch the current time from the model.
}
\seealso{
\link{dust_model_set_time}
}
