% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_model_update_pars}
\alias{dust_model_update_pars}
\title{Update parameters}
\usage{
dust_model_update_pars(model, pars)
}
\arguments{
\item{model}{A \code{dust_model} object}

\item{pars}{Parameters to set into the model.}
}
\value{
Nothing, called for side effects only
}
\description{
Update parameters used by the model.  This can be used to update a
subset of parameters that do not change the extent of the model,
and will be potentially faster than creating a new model object.
}
