% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_compare_data}
\alias{dust_system_compare_data}
\title{Compare system state against data}
\usage{
dust_system_compare_data(sys, data)
}
\arguments{
\item{sys}{A \code{dust_system} object}

\item{data}{The data to compare against. If the system is ungrouped
then \code{data} is a list with elements corresponding to whatever
your system requires.  If your system is grouped, this should be a
list with as many elements as your system has groups, with each
element corresponding to the data your system requires.}
}
\value{
A numeric vector with as many elements as your system has
groups, corresponding to the log-likelihood of the data for each
group.
}
\description{
Compare current system state against data.  This is only supported
for systems that have 'compare_data' support (i.e., the system
definition includes a \code{compare_data} method).  The current state
in the system (\link{dust_system_state}) is compared against the data
provided as \code{data}.
}
