% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_generator}
\alias{dust_system_generator}
\title{Create a system generator}
\usage{
dust_system_generator(name, time_type, env = parent.env(parent.frame()))
}
\arguments{
\item{name}{The name of the generator}

\item{time_type}{The time type (discrete or continuous).  Using
the wrong time here will lead to crashes or failure to create
the generator.}

\item{env}{The environment where the generator is defined.}
}
\value{
A \code{dust_system_generator} object
}
\description{
Create a system generator.  This function is not intended to be
used by users directly, but will be called from packages built
using \link{dust_package}
}
\examples{
# This creates the "sir" generator
dust_system_generator("sir", "discrete", asNamespace("dust2"))

# This is the same code as in "dust2:::sir", except there we find
# the correct environment automatically
dust2:::sir
}
\keyword{internal}
