% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_internals}
\alias{dust_system_internals}
\title{Fetch system internals}
\usage{
dust_system_internals(sys)
}
\arguments{
\item{sys}{A \code{dust_system} object}
}
\value{
If \code{sys} is a discrete-time system, this function returns
\code{NULL}, as no internal data is stored.  Otherwise, for a
continuous-time system we return a \code{data.frame} of statistics
with one row per particle.  Most of the columns are simple
integers or numeric values, but \code{dydt} (the current derivative
of the target function with respect to time) and \code{step_times}
(times that the solver has stopped at, if
\code{debug_record_step_times} is in \link{dust_ode_control} was set to
\code{TRUE}) will be a list of columns, each element of which is a numeric
vector.
}
\description{
Return internal data from the system.  This is intended for
debugging only, and all formats are subject to change.
}
