% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_rng_state}
\alias{dust_system_rng_state}
\alias{dust_system_set_rng_state}
\title{Fetch and set rng state}
\usage{
dust_system_rng_state(sys)

dust_system_set_rng_state(sys, rng_state)
}
\arguments{
\item{sys}{A \code{dust_system} object}

\item{rng_state}{A raw vector of random number generator state,
returned by \code{dust_system_rng_state}}
}
\value{
A raw vector, this could be quite long.
}
\description{
Fetch, and set, the random number generator (RNG) state from the
system.
}
\seealso{
You can pass the state you get back from this function as
the seed object to \code{dust_system_create} and
\code{dust_system_set_rng_state}
}
