% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_run_steps}
\alias{dust_system_run_steps}
\alias{dust_system_run_to_time}
\title{Run system}
\usage{
dust_system_run_steps(sys, steps)

dust_system_run_to_time(sys, time)
}
\arguments{
\item{sys}{A \code{dust_system} object}

\item{steps}{The number of steps to run forward}

\item{time}{Time to run to}
}
\value{
Nothing, called for side effects only
}
\description{
Run a system, advancing time and the state by repeatedly running
its \code{update} method.  You can advance a system either a fixed
(positive) number of steps, or up to a time (which must be in the
future).
}
