% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_set_time}
\alias{dust_system_set_time}
\title{Set system time}
\usage{
dust_system_set_time(sys, time)
}
\arguments{
\item{sys}{A \code{dust_system} object}

\item{time}{The time to set.  Currently this must be an
integer-like value, but in future we will allow setting to any
multiple of \code{dt}.}
}
\value{
Nothing, called for side effects only
}
\description{
Set time into the system.  This updates the time to the provided
value but does not affect the state.  You may want to call
\link{dust_system_set_state} or \link{dust_system_set_state_initial} after
calling this.
}
