% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_time}
\alias{dust_system_time}
\title{Fetch system time}
\usage{
dust_system_time(sys)
}
\arguments{
\item{sys}{A \code{dust_system} object}
}
\value{
A single numeric value
}
\description{
Fetch the current time from the system.
}
\seealso{
\link{dust_system_set_time}
}
