% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_update_pars}
\alias{dust_system_update_pars}
\title{Update parameters}
\usage{
dust_system_update_pars(sys, pars)
}
\arguments{
\item{sys}{A \code{dust_system} object}

\item{pars}{Parameters to set into the system.}
}
\value{
Nothing, called for side effects only
}
\description{
Update parameters used by the system.  This can be used to update a
subset of parameters that do not change the extent of the system,
and will be potentially faster than creating a new system object.
}
