% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-unfilter.R
\name{dust_unfilter_create}
\alias{dust_unfilter_create}
\title{Create an unfilter}
\usage{
dust_unfilter_create(
  generator,
  time_start,
  time,
  data,
  n_particles = 1,
  n_groups = 0,
  dt = 1,
  index = NULL
)
}
\arguments{
\item{generator}{A system generator object, with class
\code{dust_system_generator}.  The system must support \code{compare_data}
to be used with this function.}

\item{time_start}{The start time for the simulation - this is
typically before the first data point.  Must be an integer-like
value.}

\item{time}{A vector of times, each of which has a corresponding
entry in \code{data}.  The system will stop at each of these times to
compute the likelihood using the compare function.}

\item{data}{The data to compare against.  This must be a list with
the same length as \code{time}, each element of which corresponds to
the data required for the system.  If the system is ungrouped then
each element of \code{data} is a list with elements corresponding to
whatever your system requires.  If your system is grouped, this
should be a list with as many elements as your system has groups,
with each element corresponding to the data your system requires.
We will likely introduce a friendlier data.frame based input
soon.}

\item{n_particles}{The number of particles to run.  Typically this
is 1, but you can run with more than 1 if you want - currently
they produce the same likelihood but if you provide different
initial conditions then you would see different likelihoods.}

\item{n_groups}{Optionally, the number of parameter groups}

\item{dt}{The time step for the system, defaults to 1}

\item{index}{An optional index of states to extract.  If given,
then we subset the system state on return.  You can use this to
return fewer system states than the system ran with, to reorder
states, or to name them on exit (names present on the index will
be copied into the rownames of the returned array).}
}
\value{
A \code{dust_unfilter} object, which can be used with
\link{dust_unfilter_run}
}
\description{
Create an "unfilter" object, which can be used to compute a
deterministic likelihood following the same algorithm as the
particle filter, but limited to a single particle.  The name for
this method will change in future.
}
