% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-unfilter.R
\name{dust_unfilter_last_gradient}
\alias{dust_unfilter_last_gradient}
\title{Fetch last unfilter gradient}
\usage{
dust_unfilter_last_gradient(unfilter, index_group = NULL)
}
\arguments{
\item{unfilter}{A \code{dust_unfilter} object, created by
\link{dust_unfilter_create}}

\item{index_group}{An optional vector of group indices to run the
filter for.  You can use this to run a subset of possible
groups, though at the moment this will likely crash.}
}
\value{
A vector (if ungrouped) or a matrix (if grouped).
}
\description{
Fetch the last gradient created by running an unfilter.  This
errors if the last call to \link{dust_unfilter_run} did not use
\code{adjoint = TRUE}.  The first time you call this (after a
particular set of parameters) it will trigger running the reverse
model.
}
