% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-unfilter.R
\name{dust_unfilter_last_state}
\alias{dust_unfilter_last_state}
\title{Fetch last unfilter state}
\usage{
dust_unfilter_last_state(unfilter, index_group = NULL)
}
\arguments{
\item{unfilter}{A \code{dust_unfilter} object, created by
\link{dust_unfilter_create}}

\item{index_group}{An optional vector of group indices to run the
filter for.  You can use this to run a subset of possible
groups, once the filter is initialised (this argument must be
\code{NULL} on the \strong{first} call).}
}
\value{
An array
}
\description{
Fetch the last state created by running an unfilter.
}
