% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_unfilter_run}
\alias{dust_unfilter_run}
\title{Run unfilter}
\usage{
dust_unfilter_run(unfilter, pars = NULL, initial = NULL)
}
\arguments{
\item{unfilter}{A \code{dust_unfilter} object, created by
\link{dust_unfilter_create}}

\item{pars}{Optional parameters to run the filter with.  If not
provided, parameters are not updated}

\item{initial}{Optional initial conditions, as a matrix (state x
particle) or 3d array (state x particle x group).  If not
provided, the model initial conditions are used.}
}
\value{
A vector of likelihood values, with as many elements as
there are groups.
}
\description{
Run unfilter
}
