% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-filter.R
\name{dust_filter_create}
\alias{dust_filter_create}
\title{Create a particle filter}
\usage{
dust_filter_create(
  generator,
  time_start,
  data,
  n_particles,
  n_groups = NULL,
  dt = 1,
  index_state = NULL,
  n_threads = 1,
  preserve_group_dimension = FALSE,
  seed = NULL
)
}
\arguments{
\item{generator}{A system generator object, with class
\code{dust_system_generator}.  The system must support \code{compare_data}
to be used with this function.}

\item{time_start}{The start time for the simulation - this is
typically before the first data point.  Must be an integer-like
value.}

\item{data}{The data to fit to.  This can be a \code{data.frame}, in
which case it will be passed into \link{dust_filter_data} for
validation, or it can be a \link{dust_filter_data}-augmented
data.frame.  The times for comparison will be taken from this,
and \code{time_start} must be no later than than the earliest time.}

\item{n_particles}{The number of particles to run.  Larger numbers
will give lower variance in the likelihood estimate but run more
slowly.}

\item{n_groups}{The number of parameter groups.  If \code{NULL}, this
will be taken from \code{data}.  If given, then the number of groups
in \code{data} will be checked against this number.}

\item{dt}{The time step for discrete time systems, defaults to 1
if not given.  It is an error to provide a non-NULL argument
with continuous-time systems.}

\item{index_state}{An optional index of states to extract.  If
given, then we subset the system state on return.  You can use
this to return fewer system states than the system ran with, to
reorder states, or to name them on exit (names present on the
index will be copied into the rownames of the returned array).}

\item{n_threads}{Integer, the number of threads to use in
parallelisable calculations.  See Details.}

\item{preserve_group_dimension}{Logical, indicating if state and
output from the system should preserve the group dimension in
the case where a single group is run.  In the case where more
than one group is run, this argument has no effect as the
dimension is always preserved.}

\item{seed}{Optionally, a seed.  Otherwise we respond to R's RNG seed on
initialisation.}
}
\value{
A \code{dust_likelihood} object, which can be used with
\link{dust_likelihood_run}
}
\description{
Create a particle filter object
}
