% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_filter_last_history}
\alias{dust_filter_last_history}
\title{Fetch last filter history}
\usage{
dust_filter_last_history(filter)
}
\arguments{
\item{filter}{A \code{dust_filter} object, created by
\link{dust_filter_create}}
}
\value{
An array
}
\description{
Fetch the last history created by running a filter.  This
errors if the last call to \link{dust_filter_run} did not use
\code{save_history = TRUE}.
}
