% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-likelihood.R
\name{dust_likelihood_run}
\alias{dust_likelihood_run}
\title{Compute likelihood}
\usage{
dust_likelihood_run(
  obj,
  pars,
  initial = NULL,
  save_trajectories = FALSE,
  save_snapshots = NULL,
  adjoint = NULL,
  index_state = NULL,
  index_group = NULL
)
}
\arguments{
\item{obj}{A \code{dust_filter} object, created by
\link{dust_filter_create} or a \code{dust_unfilter} object created by
\link{dust_unfilter_create}}

\item{pars}{Optional parameters to compute the likelihood with.
If not provided, parameters are not updated}

\item{initial}{Optional initial conditions, as a matrix (state x
particle) or 3d array (state x particle x group).  If not
provided, the system initial conditions are used.}

\item{save_trajectories}{Logical, indicating if the trajectories
from the simulation history should be saved while the simulation
runs; this has a small overhead in runtime and in memory.
Particle trajectories will be saved at each time for which you
have data.  If \code{index_state} is non-\code{NULL}, the trajectories are
limited to these states.}

\item{save_snapshots}{Optional vector of times at which we should
record snapshots of the entire state of the system.}

\item{adjoint}{Optional logical, indicating if we should enable
adjoint history saving. This is enabled by default if your model
has an adjoint, but can be disabled or enabled even when your
model does not support adjoints! But if you don't actually have
an adjoint you will not be able to compute gradients.  This has
no effect for stochastic models.}

\item{index_state}{An optional vector of state indices to save
where \code{save_trajectories} is \code{TRUE}.  If omitted all state is saved
and if \code{save_trajectories = FALSE} this has no effect.}

\item{index_group}{An optional vector of group indices to run the
calculation for.  You can use this to run a subset of possible
groups, once \code{obj} is initialised (this argument must be \code{NULL}
on the \strong{first} call).}
}
\value{
A vector of likelihood values, with as many elements as
there are groups.
}
\description{
Compute a log likelihood based on a dynamical model, either from
particle filter (created with \link{dust_filter_create}) or a
deterministic model (created with \link{dust_unfilter_create}).
}
