% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_system_set_state}
\alias{dust_system_set_state}
\title{Set system state}
\usage{
dust_system_set_state(sys, state)
}
\arguments{
\item{sys}{A \code{dust_system} object}

\item{state}{A matrix or array of state.  If ungrouped, the
dimension order expected is state x particle.  If grouped the
order is state x particle x group.}
}
\value{
Nothing, called for side effects only
}
\description{
Set system state.  Takes a multidimensional array (2- or 3d
depending on if the system is grouped or not).  Dimensions of
length 1 will be recycled as appropriate.
}
