% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-unfilter.R
\name{dust_unfilter_last_history}
\alias{dust_unfilter_last_history}
\title{Fetch last unfilter history}
\usage{
dust_unfilter_last_history(unfilter)
}
\arguments{
\item{unfilter}{A \code{dust_unfilter} object, created by
\link{dust_unfilter_create}}
}
\value{
An array
}
\description{
Fetch the last history created by running an unfilter.  This
errors if the last call to \link{dust_unfilter_run} did not use
\code{save_history = TRUE}.
}
