% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dust_unfilter_run}
\alias{dust_unfilter_run}
\title{Run unfilter}
\usage{
dust_unfilter_run(unfilter, pars = NULL, initial = NULL, save_history = FALSE)
}
\arguments{
\item{unfilter}{A \code{dust_unfilter} object, created by
\link{dust_unfilter_create}}

\item{pars}{Optional parameters to run the filter with.  If not
provided, parameters are not updated}

\item{initial}{Optional initial conditions, as a matrix (state x
particle) or 3d array (state x particle x group).  If not
provided, the model initial conditions are used.}

\item{save_history}{Logical, indicating if the simulation history
should be saved while the simulation runs; this has a small
overhead in runtime and in memory.  History (particle
trajectories) will be saved at each time in the filter.  If the
filter was constructed using a non-\code{NULL} \code{index} parameter, the
the history is restricted to these states.}
}
\value{
A vector of likelihood values, with as many elements as
there are groups.
}
\description{
Run unfilter
}
